+++ app/(routes)/guide/shorts-viewer/page.tsx
@@ -1,0 +1,5 @@
+import ShortsViewer from './_components/shorts-viewer-container';
+
+export default function Page() {
+  return <ShortsViewer />;
+}
